ignore DeleteOrphanPvc on cluster creation (#218)
* ignore DeleteOrphanPvc on cluster creation

* refactor condition for deleteorphan and generation

+++ controllers/druid/handler.go
@@ -272,12 +272,14 @@ func deployDruidCluster(sdk client.Client, m *v1alpha1.Druid) error {
 		}
 	}
 
-	if m.Spec.DeleteOrphanPvc {
+	// Ignore on cluster creation
+	if m.Generation > 1 && m.Spec.DeleteOrphanPvc {
 		if err := deleteOrphanPVC(sdk, m); err != nil {
 			e := fmt.Errorf("Error in deleteOrphanPVC due to [%s]", err.Error())
-			sendEvent(sdk, m, v1.EventTypeWarning, "LIST_FAIL", e.Error())
+			sendEvent(sdk, m, v1.EventTypeWarning, DruidNodeDeleteFail, e.Error())
 			logger.Error(e, e.Error(), "name", m.Name, "namespace", m.Namespace)
 		}
+
 	}
 
 	//update status and delete unwanted resources
